Tests/LibWeb/Text/expected/HTML/maxed-out-dimension-value.txt

Image height: 17895700
